+++ packages/backend/src/misc/is-hidden-or-visibility-modified.ts
