+++ packages/backend/src/server/api/endpoints/notes/search.ts
@@ -8,7 +8,7 @@ import { Endpoint } from '@/server/api/endpoint-base.js';
 import { SearchService } from '@/core/SearchService.js';
 import { NoteEntityService } from '@/core/entities/NoteEntityService.js';
 import { RoleService } from '@/core/RoleService.js';
-import { isMustRemove } from '@/misc/is-hidden-or-visibilityModified.js';
+import { isMustRemove } from '@/misc/is-hidden-or-visibility-modified.js';
 import { ApiError } from '../../error.js';
 
 export const meta = {
