+++ packages/backend/src/server/api/stream/channels/antenna.ts
@@ -8,7 +8,7 @@ import { NoteEntityService } from '@/core/entities/NoteEntityService.js';
 import { bindThis } from '@/decorators.js';
 import type { GlobalEvents } from '@/core/GlobalEventService.js';
 import type { JsonObject } from '@/misc/json-value.js';
-import { isMustRemove } from '@/misc/is-hidden-or-visibilityModified.js';
+import { isMustRemove } from '@/misc/is-hidden-or-visibility-modified.js';
 import Channel, { type MiChannelService } from '../channel.js';
 
 class AntennaChannel extends Channel {
